+++ cmd/shopts/main.go
@@ -1,6 +1,7 @@
 package main
 
 import (
+	"errors"
 	"fmt"
 	"os"
 
@@ -15,8 +16,12 @@ func main() {
 		fmt.Println(version)
 		return
 	}
-	if err := shopts.Run(os.Args, os.Stdout); err != nil {
+	if err := shopts.Run(os.Args, os.Stdout, os.Stderr); err != nil {
 		fmt.Fprintln(os.Stderr, err)
+		var exitErr *shopts.ExitError
+		if errors.As(err, &exitErr) {
+			os.Exit(exitErr.Code)
+		}
 		os.Exit(1)
 	}
 }